refactor: separate map component
