+++ apps/web/src/app/layout.tsx
@@ -4,6 +4,8 @@ import './globals.css'
 import localFont from 'next/font/local'
 import { useState } from 'react'
 import { QueryClientProvider, QueryClient } from '@tanstack/react-query'
+import { KAKAO_API_KEY } from '@/config'
+import Script from 'next/script'
 
 const geistSans = localFont({
     src: './fonts/GeistVF.woff',
@@ -15,6 +17,8 @@ const geistMono = localFont({
     variable: '--font-geist-mono',
     weight: '100 900',
 })
+const KAKAO_SDK_URL = `https://dapi.kakao.com/v2/maps/sdk.js?appkey=${KAKAO_API_KEY}&autoload=false`
+
 export default function RootLayout({ children }: Readonly<{ children: React.ReactNode }>) {
     const [queryClient] = useState(
         () =>
@@ -35,6 +39,7 @@ export default function RootLayout({ children }: Readonly<{ children: React.Reac
         <html lang="en">
             <body className={`${geistSans.variable} ${geistMono.variable} antialiased`}>
                 <QueryClientProvider client={queryClient}>{children}</QueryClientProvider>
+                <Script src={KAKAO_SDK_URL} strategy="beforeInteractive" />
             </body>
         </html>
     )
